Added read me and changed file dir
^
