+++ .gitignore
@@ -1,0 +1,2 @@
+/ESD_Assignment/nbproject/private/
+/ESDSet1/nbproject/private/
