--- ESD Pet Shop/nbproject/private/private.properties
@@ -1,5 +1,0 @@
-deploy.ant.properties.file=C:\\Users\\Edwin\\AppData\\Roaming\\NetBeans\\8.1\\tomcat80.properties
-j2ee.server.domain=C:/Users/Edwin/AppData/Roaming/NetBeans/8.1/apache-tomcat-8.0.27.0_base
-j2ee.server.home=C:/Program Files/Apache Software Foundation/Apache Tomcat 8.0.27
-j2ee.server.instance=tomcat80:home=C:\\Program Files\\Apache Software Foundation\\Apache Tomcat 8.0.27:base=apache-tomcat-8.0.27.0_base
-user.properties.file=C:\\Users\\Edwin\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
