--- ESD Pet Shop/nbproject/private/private.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/Edwin/Desktop/ESD/ESD%20Pet%20Shop/src/java/model/Pets.java</file>
-            <file>file:/C:/Users/Edwin/Desktop/ESD/ESD%20Pet%20Shop/web/WEB-INF/docs/subview.jsp</file>
-            <file>file:/C:/Users/Edwin/Desktop/ESD/ESD%20Pet%20Shop/src/java/controller/Front.java</file>
-            <file>file:/C:/Users/Edwin/Desktop/ESD/ESD%20Pet%20Shop/web/WEB-INF/docs/main.jsp</file>
-        </group>
-    </open-files>
-</project-private>
