--- ESD Pet Shop/nbproject/project.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>org.netbeans.modules.web.project</type>
-    <configuration>
-        <data xmlns="http://www.netbeans.org/ns/web-project/3">
-            <name>ESD_Assignment</name>
-            <minimum-ant-version>1.6.5</minimum-ant-version>
-            <web-module-libraries/>
-            <web-module-additional-libraries/>
-            <source-roots>
-                <root id="src.dir" name="Source Packages"/>
-            </source-roots>
-            <test-roots>
-                <root id="test.src.dir" name="Test Packages"/>
-            </test-roots>
-        </data>
-    </configuration>
-</project>
