--- ESD Pet Shop/src/java/model/Subscription.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Subscription {
-    public static String hello() {
-        return "Hello there from Pets!";
-    }
-}
