--- ESD Pet Shop/web/META-INF/context.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Context path="/ESD_Assignment"/>
