--- ESD Pet Shop/web/WEB-INF/docs/adminPage.jsp
@@ -1,4 +1,0 @@
-<%@page import="model.Pets"%>
-<h1>Sub-View</h1>
-
-<%= Subscription.hello() %>
