+++ ESD_Assignment/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=6e2ebe27
+build.xml.script.CRC32=76b16d80
+build.xml.stylesheet.CRC32=651128d4@1.75.1.1
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=6e2ebe27
+nbproject/build-impl.xml.script.CRC32=0676534c
+nbproject/build-impl.xml.stylesheet.CRC32=99ea4b56@1.75.1.1
