+++ ESD_Assignment/src/java/model/Subscription.java
@@ -1,0 +1,7 @@
+package model;
+
+public class Subscription {
+    public static String hello() {
+        return "Hello there from Pets!";
+    }
+}
