+++ ESD_Assignment/web/META-INF/context.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Context path="/ESD_Assignment"/>
