+++ ESD_Assignment/web/WEB-INF/docs/logInPage.jsp
@@ -1,0 +1,4 @@
+<%@page import="model.Pets"%>
+<h1>Sub-View</h1>
+
+<%= Subscription.hello() %>
